mock, runtime: Create mock ethereum adapter and use in runtime test
Preparing to add ethereum_adapter to the Subgraph registrar.
